# static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OpenID Federation Resolver",
    "version": "1.0.0",
    "description": "Operations and federation APIs for this resolver instance."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "MetricsBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Required for GET /metrics only when METRICS_TOKEN is set."
      },
      "OperatorBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "API_KEY or a PAT (oidf_pat_…) for /admin/v1, cache POST/DELETE, and configuration. Also accepted as X-API-Key. GET /api/v1 routes stay public. Minting PATs requires ENV API_KEY when it is set."
      },
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "X-API-Key: API_KEY or PAT for operator routes, or TA_API_KEY for trust-anchor register/unregister."
      },
      "TABearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "TA_API_KEY for POST /register-trust-anchor and DELETE registered-trust-anchors. TA_API_TOKEN is an alias. API_KEY is not accepted."
      }
    },
...
  },
  "paths": {
...
    "/admin/v1": {
      "get": {
        "tags": ["Administration"],
        "summary": "Node document (draft-kodden-oidfed-admin-00)",
        "description": "Identity, roles (resolver), and capabilities. Control-plane discovery for OIDF Admin. Requires API_KEY when set. This node does not implement subordinates or trust marks.",
        "security": [{"OperatorBearer": []}, {"ApiKey": []}],
        "responses": {
          "200": { "description": "Node document" },
          "401": { "description": "Missing or invalid API_KEY" }
        }
      }
    },
    "/admin/v1/configuration": {
      "get": {
        "tags": ["Administration"],
        "summary": "Read Entity Configuration admin document",
        "security": [{"OperatorBearer": []}, {"ApiKey": []}],
        "responses": { "200": { "description": "Configuration document with ETag" } }
      },
      "put": {
        "tags": ["Administration"],
        "summary": "Replace Entity Configuration admin document",
        "security": [{"OperatorBearer": []}, {"ApiKey": []}],
        "responses": { "200": { "description": "Updated document" } }
      },
      "patch": {
        "tags": ["Administration"],
        "summary": "JSON Merge Patch Entity Configuration (RFC 7396)",
        "security": [{"OperatorBearer": []}, {"ApiKey": []}],
        "responses": { "200": { "description": "Updated document" } }
      }
    },
    "/admin/v1/configuration/statement": {
      "get": {
        "tags": ["Administration"],
        "summary": "Published Entity Configuration JWT",
        "description": "Byte-for-byte the document at /.well-known/openid-federation.",
        "security": [{"OperatorBearer": []}, {"ApiKey": []}],
        "responses": { "200": { "description": "application/entity-statement+jwt" } }
      }
    },
    "/admin/v1/keys": {
      "get": {
        "tags": ["Administration"],
        "summary": "List Federation Entity Keys",
        "security": [{"OperatorBearer": []}, {"ApiKey": []}],
        "responses": { "200": { "description": "Paged Key Documents" } }
      },
      "post": {
        "tags": ["Administration"],
        "summary": "Generate a Federation Entity Key",
        "security": [{"OperatorBearer": []}, {"ApiKey": []}],
        "responses": { "201": { "description": "Created Key Document" } }
      }
    },
    "/admin/v1/keys/{kid}": {
      "get": {
        "tags": ["Administration"],
        "summary": "Read a Federation Entity Key",
        "security": [{"OperatorBearer": []}, {"ApiKey": []}],
        "parameters": [{ "name": "kid", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Key Document" } }
      },
      "delete": {
        "tags": ["Administration"],
        "summary": "Revoke a Federation Entity Key",
        "security": [{"OperatorBearer": []}, {"ApiKey": []}],
        "parameters": [{ "name": "kid", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "204": { "description": "Revoked" },
          "409": { "description": "last_signing_key" }
        }
      }
    },
    "/admin/v1/keys/{kid}/rotate": {
      "post": {
        "tags": ["Administration"],
        "summary": "Roll over a Federation Entity Key",
        "security": [{"OperatorBearer": []}, {"ApiKey": []}],
        "parameters": [{ "name": "kid", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "201": { "description": "New Key Document" } }
      }
    },
    "/api/v1/tokens": {
      "get": {
        "tags": ["Tokens"],
        "summary": "List personal access tokens",
        "description": "Secret-free PAT metadata. When API_KEY is set, only the ENV key may list or mint tokens (a PAT cannot manage other PATs).",
        "security": [{"OperatorBearer": []}, {"ApiKey": []}],
        "responses": {
          "200": { "description": "Token list" },
          "401": { "description": "Missing or invalid credentials" },
          "403": { "description": "PAT presented; ENV API_KEY required to manage tokens" }
        }
      },
      "post": {
        "tags": ["Tokens"],
        "summary": "Create a personal access token",
        "description": "Returns the plaintext secret once. Prefix oidf_pat_. Body: { name, ttl_days }.",
        "security": [{"OperatorBearer": []}, {"ApiKey": []}],
        "responses": {
          "201": { "description": "Created; token field is the secret" },
          "401": { "description": "Missing or invalid credentials" },
          "403": { "description": "PAT presented; ENV API_KEY required to manage tokens" }
        }
      }
    },
    "/api/v1/tokens/{id}": {
      "get": {
        "tags": ["Tokens"],
        "summary": "Read a personal access token",
        "security": [{"OperatorBearer": []}, {"ApiKey": []}],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Token metadata" },
          "404": { "description": "Unknown token" }
        }
      },
      "delete": {
        "tags": ["Tokens"],
        "summary": "Revoke a personal access token",
        "security": [{"OperatorBearer": []}, {"ApiKey": []}],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Revoked" },
          "404": { "description": "Unknown token" }
        }
      }
    }
  }
}